pyproject.toml
[project]
name = "stock-insight-writer"
version = "0.1.0"
description = "AI-powered daily stock investment post writer using LLMs and market data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.50.0",
    "yfinance>=0.2.50",
    "httpx>=0.28.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-slugify>=8.0",
    "tenacity>=9.0",
    "typer>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "debugpy>=1.8.0",
]

[project.scripts]
stock-writer = "stock_insight_writer.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_insight_writer"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]  # Typer requires function calls in default arguments

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
